src/requirements.txt
mcp[cli]>=1.7
netmiko>=4.4
python-dotenv>=1.0
pyyaml>=6.0
